apps/contextractor-standalone/pyproject.toml
[project]
name = "contextractor-standalone"
version = "0.1.0"
description = "Standalone CLI for web content extraction"
requires-python = ">=3.12"
dependencies = [
    "crawlee[playwright]>=0.4.0",
    "contextractor-engine",
    "typer>=0.15.0",
    "pyyaml>=6.0",
    "browserforge<1.2.4",
]

[project.scripts]
contextractor = "contextractor_cli.main:app"

[tool.uv.sources]
contextractor-engine = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/contextractor_cli"]
